Moved Price and Quantity case classes out of package object.
